Install additional dependencies, tweak settings
Make some additional changes that are required for the Genesis
test suite.  Issue we are facing is that _currently_ it has to
be run _from_ a Zimbra node, so we have to add test dependencies
on that node.  In addition, we have to configure timezone
accordingly.

NOTE: In the Dockerfile have configured tzdata for UTC.  Will be updating
      the docker-compose init script to reconfigure tzdata based on
      environment variable settings.